refactor to sequences
