+++ web-client/src/presenter/actions/openPdfPreviewTabAction.js
@@ -1,0 +1,141 @@
+const renderPdfViewer = pdfData => {
+  const base64 = pdfData.replace(/[^,]+,/, '');
+  return `
+  <html>
+    <head>
+      <div>
+        <button id="prev">Previous</button>
+        <button id="next">Next</button>
+        &nbsp; &nbsp;
+        <span>Page: <span id="page_num"></span> / <span id="page_count"></span></span>
+      </div>
+      
+      <canvas id="the-canvas"></canvas>
+      <script src="http://localhost:1234/pdf.min.js"></script>
+
+      <style>
+      </style>
+    </head>
+
+    <body>
+      <script>
+        var pdfData = atob("${base64}");
+        // Loaded via <script> tag, create shortcut to access PDF.js exports.
+        var pdfjsLib = window['pdfjs-dist/build/pdf'];
+
+        // The workerSrc property shall be specified.
+        pdfjsLib.GlobalWorkerOptions.workerSrc = '/pdf.worker.js';
+
+        var pdfDoc = null,
+            pageNum = 1,
+            pageRendering = false,
+            pageNumPending = null,
+            scale = 0.8,
+            canvas = document.getElementById('the-canvas'),
+            ctx = canvas.getContext('2d');
+
+        /**
+         * Get page info from document, resize canvas accordingly, and render page.
+         * @param num Page number.
+         */
+        function renderPage(num) {
+          pageRendering = true;
+          // Using promise to fetch the page
+          pdfDoc.getPage(num).then(function(page) {
+            var viewport = page.getViewport({scale: scale});
+            canvas.height = viewport.height;
+            canvas.width = viewport.width;
+
+            // Render PDF page into canvas context
+            var renderContext = {
+              canvasContext: ctx,
+              viewport: viewport
+            };
+            var renderTask = page.render(renderContext);
+
+            // Wait for rendering to finish
+            renderTask.promise.then(function() {
+              pageRendering = false;
+              if (pageNumPending !== null) {
+                // New page rendering is pending
+                renderPage(pageNumPending);
+                pageNumPending = null;
+              }
+            });
+          });
+
+          // Update page counters
+          document.getElementById('page_num').textContent = num;
+        }
+
+        /**
+         * If another page rendering in progress, waits until the rendering is
+         * finised. Otherwise, executes rendering immediately.
+         */
+        function queueRenderPage(num) {
+          if (pageRendering) {
+            pageNumPending = num;
+          } else {
+            renderPage(num);
+          }
+        }
+
+        /**
+         * Displays previous page.
+         */
+        function onPrevPage() {
+          if (pageNum <= 1) {
+            return;
+          }
+          pageNum--;
+          queueRenderPage(pageNum);
+        }
+        document.getElementById('prev').addEventListener('click', onPrevPage);
+
+        /**
+         * Displays next page.
+         */
+        function onNextPage() {
+          if (pageNum >= pdfDoc.numPages) {
+            return;
+          }
+          pageNum++;
+          queueRenderPage(pageNum);
+        }
+        document.getElementById('next').addEventListener('click', onNextPage);
+
+        /**
+         * Asynchronously downloads PDF.
+         */
+        pdfjsLib.getDocument({ data: pdfData}).promise.then(function(pdfDoc_) {
+          pdfDoc = pdfDoc_;
+          document.getElementById('page_count').textContent = pdfDoc.numPages;
+
+          // Initial/first page rendering
+          renderPage(pageNum);
+        });
+      </script>
+    </body>
+  </html>
+  `;
+};
+
+/**
+ * opens a new tab and uses pdf.js to preview a pdf file stored locally in the store
+ *
+ * @param {object} providers the providers object
+ * @param {object} providers.router the riot.router object that is used for changing the route
+ * @param {Function} providers.get the cerebral get function used for getting the path to navigate to in state.path
+ */
+export const openPdfPreviewTabAction = ({ props }) => {
+  var reader = new FileReader();
+  reader.readAsDataURL(props.file);
+  reader.onload = function() {
+    const pdfData = reader.result;
+    const previewWindow = window.open();
+    previewWindow.document.write(renderPdfViewer(pdfData));
+  };
+  reader.onerror = function(error) {
+    console.log('error: ', error);
+  };
+};
