+++ web-client/src/presenter/sequences/openPdfPreviewTabSequence.js
@@ -1,0 +1,3 @@
+import { openPdfPreviewTabAction } from '../actions/openPdfPreviewTabAction';
+
+export const openPdfPreviewTabSequence = [openPdfPreviewTabAction];
